backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
youtube-transcript-api>=0.6.0
google-generativeai>=0.3.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
yt-dlp>=2023.12.30
redis>=5.0.0

# Enhanced multimodal capabilities for production improvements
torch>=2.0.0
torchvision>=0.15.0
clip-by-openai==1.0
transformers>=4.30.0
